made Address finite
Basically removed all `valid_node` related parts as a result.
Note that although in the future, the set of valid nodes may
change due to reconfiguration, the current definition of
`valid_node` will remain useless since in that case, we will have
to develop a `valid_node` which accepts time as an argument.