separate `t0` and `is_byz` from the `Address` module
The motivation is that `t0` should be related with the protocol
implementation, and `is_byz` should essentially be a universally
quantified function. They should not be put into `Address`, which
is no more than a datatype module.
Also, required stdpp (should be fine).
Also, parameterized the threshold of the threshold signature
scheme module (originally it was `t0`, which is not precise).
Though for ABC, the threshold can be set to `t0` (implemented
by using the `with Definition` mechanism of Coq module).